refactor: Improve habit form view namespaces
